Item.align_for_transform

+++ janim/utils/iterables.py
@@ -93,7 +93,7 @@ def resize_array(nparray: np.ndarray, length: int) -> np.ndarray:
 
 def resize_preserving_order(nparray: np.ndarray, length: int) -> np.ndarray:
     if len(nparray) == 0:
-        return np.zeros((length, *nparray.shape[1:]))
+        return np.zeros((length, *nparray.shape[1:]), dtype=nparray.dtype)
     if len(nparray) == length:
         return nparray
     indices = np.arange(length) * len(nparray) // length
@@ -104,13 +104,13 @@ def resize_with_interpolation(nparray: np.ndarray, length: int) -> np.ndarray:
     if len(nparray) == length:
         return nparray
     if length == 0:
-        return np.zeros((0, *nparray.shape[1:]))
+        return np.zeros((0, *nparray.shape[1:]), dtype=nparray.dtype)
     cont_indices = np.linspace(0, len(nparray) - 1, length)
     return np.array([
         (1 - a) * nparray[lh] + a * nparray[rh]
         for ci in cont_indices
         for lh, rh, a in [(int(ci), int(np.ceil(ci)), ci % 1)]
-    ])
+    ], dtype=nparray.dtype)
 
 
 def make_even(
